Created a hill page and moved non hill logic out of the hill component
