+++ client/src/pages/HillPage/interfaces/IHillPageProps.ts
@@ -1,0 +1,8 @@
+
+import { Socket } from "socket.io-client/build/socket";
+import { DefaultEventsMap } from "socket.io-client/build/typed-events";
+
+export interface IHillPageProps {
+    id: string,
+    socket: Socket<DefaultEventsMap, DefaultEventsMap>
+}
