+++ client/src/pages/HillPage/interfaces/IHillPageState.ts
@@ -1,0 +1,11 @@
+import { IHill } from "../../../components/Hill/interfaces/IHill";
+import { IMarker } from "../../../components/Marker/interfaces/IMarker";
+import { IMarkerData } from "../../../util/IMarkerData";
+
+export interface IHillPageState {
+    hill: IHill,
+    markers: IMarker[],
+    activeModal: string,
+    selectedMarker: IMarkerData,
+    isMarkerClick: boolean
+}
